backend/requirements.txt
# SafeBrowse Backend - Pinned Dependencies
# Generated: 2024-12-17

# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0

# LLM Integration
openai==1.58.1

# Environment & Config
python-dotenv==1.0.1

# HTML Parsing
beautifulsoup4==4.12.3
lxml==5.3.0

# Testing
pytest==8.3.4
httpx==0.28.1
pytest-asyncio==0.25.0

# Optional: Document scanning
# pytesseract==0.3.13  # Requires Tesseract OCR
# pymupdf==1.25.1      # PDF support
# pillow==11.0.0       # Image support

# Optional: ML safety model
# torch>=2.0.0
# transformers>=4.30.0

# Database & Auth
sqlmodel==0.0.14
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-multipart==0.0.6
